# Java/AssetsExchange/exchangeassets/trading-engine/openapi.yml
openapi: 3.0.0
info:
  title: Trading Engine API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /orders:
    get:
      summary: Get all orders
      operationId: getAllOrders
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
    post:
      summary: Submit an order
      operationId: submitOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmittedOrder'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderSubmissionResponse'
  /orders/{id}:
    delete:
      summary: Delete a not-fulfilled order
      operationId: cancelOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
components:
  schemas:
    Order:
      type: object
      properties:
        id:
          type: string
        side:
          type: string
          enum: [BUY, SELL]
        type:
          type: string
          enum: [LIMIT]
        numContracts:
          type: integer
        limitPrice:
          type: integer
        currency:
          type: string
          enum: [USD, EUR, GBP, JPY]
        remainingContracts:
          type: integer
        status:
          type: string
          enum: [Unfulfilled, PartiallyFilled, Fulfilled]
        matches:
          type: array
          items: {}
        notFulfilled:
          type: boolean
        fulfilled:
          type: boolean
      required:
        - id
        - side
        - type
        - numContracts
        - limitPrice
        - currency
        - remainingContracts
        - status
    SubmittedOrder:
      type: object
      properties:
        side:
          type: string
          enum: [BUY, SELL]
        type:
          type: string
          enum: [LIMIT]
        numContracts:
          type: integer
        limitPrice:
          type: integer
        currency:
          type: string
          enum: [USD, EUR, GBP, JPY]
    OrderSubmissionResponse:
      type: object
      properties:
        order_id:
          type: string
        matches:
          type: array
          items: {}